@@ -12,28 +12,30 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^8.0.0",
+    "@angular/animations": "^10.0.8",
     "@angular/cdk": "^8.0.0",
-    "@angular/common": "^8.0.0",
-    "@angular/compiler": "^8.0.0",
-    "@angular/core": "^8.0.0",
-    "@angular/forms": "^8.0.0",
+    "@angular/common": "^10.0.8",
+    "@angular/compiler": "^10.0.8",
+    "@angular/core": "^10.0.8",
+    "@angular/forms": "^10.0.8",
     "@angular/material": "^8.0.0",
-    "@angular/platform-browser": "^8.0.0",
-    "@angular/platform-browser-dynamic": "^8.0.0",
-    "@angular/router": "^8.0.0",
-    "angular-font-awesome": "^3.1.2",
+    "@angular/platform-browser": "^10.0.8",
+    "@angular/platform-browser-dynamic": "^10.0.8",
+    "@angular/router": "^10.0.8",
+    "@fortawesome/angular-fontawesome": "^0.7.0",
+    "@fortawesome/fontawesome-svg-core": "^1.2.30",
+    "@fortawesome/free-solid-svg-icons": "^5.14.0",
     "core-js": "^2.6.0",
     "font-awesome": "^4.7.0",
     "rxjs": "^6.3.3",
     "rxjs-compat": "^6.3.3",
-    "zone.js": "~0.9.1"
+    "zone.js": "~0.10.3"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^0.1000.5",
     "@angular/cli": "^10.0.5",
     "@angular/compiler-cli": "^10.0.8",
-    "@angular/language-service": "^8.0.0",
+    "@angular/language-service": "^10.0.8",
     "@types/jasmine": "~3.3.0",
     "@types/jasminewd2": "~2.0.6",
     "@types/node": "~10.12.12",
@@ -49,6 +51,6 @@
     "protractor": "^7.0.0",
     "ts-node": "~7.0.1",
     "tslint": "~5.11.0",
-    "typescript": "~3.4.5"
+    "typescript": "~3.9.7"
   }
 }
