@@ -1,12 +1,12 @@
-import {NgModule} from '@angular/core';
-import {CommonModule} from '@angular/common';
-import {ChuckJokesComponent} from './chuck-jokes/chuck-jokes.component';
-import {DadJokesComponent} from './dad-jokes/dad-jokes.component';
-import { MatButtonModule } from '@angular/material/button';
-import { MatCardModule } from '@angular/material/card';
-import { MatExpansionModule } from '@angular/material/expansion';
 import {HttpClientModule} from '@angular/common/http';
-import {AngularFontAwesomeModule} from 'angular-font-awesome';
+import { NgModule } from "@angular/core";
+import { CommonModule } from "@angular/common";
+import {FontAwesomeModule} from '@fortawesome/angular-fontawesome';
+import { ChuckJokesComponent } from "./chuck-jokes/chuck-jokes.component";
+import { DadJokesComponent } from "./dad-jokes/dad-jokes.component";
+import { MatButtonModule } from "@angular/material/button";
+import { MatCardModule } from "@angular/material/card";
+import { MatExpansionModule } from "@angular/material/expansion";
 
 @NgModule({
   imports: [
@@ -14,13 +14,10 @@ import {AngularFontAwesomeModule} from 'angular-font-awesome';
     MatButtonModule,
     MatExpansionModule,
     MatCardModule,
-    AngularFontAwesomeModule,
-    HttpClientModule
+    FontAwesomeModule,
+    HttpClientModule,
   ],
-  exports: [
-    DadJokesComponent,
-    ChuckJokesComponent
-  ],
-  declarations: [ChuckJokesComponent, DadJokesComponent]
+  exports: [DadJokesComponent, ChuckJokesComponent],
+  declarations: [ChuckJokesComponent, DadJokesComponent],
 })
-export class JokesModule { }
+export class JokesModule {}
